결제 준비 로직 구현 - 일반 결제 저장 처리 및 구독 결제 시 tbl_subscribe 생성, tbl_subscribe_round 회차(1회차) 자동 생성 및 결제(tbl_payment)와 연결

+++ src/main/java/com/imchobo/sayren_back/domain/subscribe/entity/Subscribe.java
@@ -45,12 +45,12 @@ public class Subscribe extends BaseEntity {
   // 월 렌탈료 스냅샷
   // NOT NULL, 기본값 0
   @Column(name = "monthly_fee_snapshot", nullable = false)
-  private Integer monthlyFeeSnapshot;
+  private Long monthlyFeeSnapshot;
 
   // 보증금 스냅샷
   // NOT NULL, 기본값 0
   @Column(name = "deposit_snapshot", nullable = false)
-  private Integer depositSnapshot;
+  private Long depositSnapshot;
 
   // 구독 시작일
   // NOT NULL
